put coriolis and topography fields into state.fields correctly
added Sum diagnostic field so that we can calculate D+b diagnostic for shallow water - this made me realise some changes were required to the Difference diagnostic field so that Sum and Difference can be added to the diagnostic fields list that is passed in to state. The problem previously was that the fields to be added/subtracted had not yet been made so now we just pass in the name of the fields so that they can be extracted from state.fields when required.