Fix tests that previously used json to use new netcdf format
Add total diagnostic to list of available diagnostics and move the exception (it only works for scalar fields) to the total compute method as it makes it easier to loop over the diagnostic fields. If anyone can think of a better way to do this that'd be great...

Move the test_sk_nonlinear test to test_checkpointing as this is what it actually does.

Fix checkpointing and netcdf file output.