Test for subcycling.
This test simulates the propagation of waves from an initial Gaussian hump in the depth field by solving the shallow water equations on an f-plane. This simulation fails at this resolution and timestep without the subcycling. All the test does is check that the subcycled simululation runs past the point that the non-subcycled simulation would blow up.