Checkpoint bug bypassed
To avoid the h5 viewer error that appears after repeatedly opening a checkpoint file: move Dumbcheckpoint to setup_dump, so that it is only created once. Note that we can't use the "with" statement in dump in this case since it will close the checkpoint (in which case we would have to open it again each time,  leading to the original error).