Advection schemes now label the equation terms to be explicit or implicit.
Some tidying in advection tests and a new tracer blob setup in conftest.

Stop being lazy with diffusion test and actually compare to the exact solution.

Move diffusion terms on to the lhs and update backward Euler advection scheme to account for this.

New helper function has_labels.

New equation class AdvectionDiffusion and test.

Some features of the Timestepping class are tested in these tests. Namely that prescribed fields work (test_IP_diffusion calculates the exact solution as a prescribed field) and that the labels can be used to apply different schemes to different parts of the equation (test_advection_diffusion uses SSPRK3 for advection and backward Euler for diffusion).